Separate API parsing from transport
